requirements.txt
attrs
bibtexparser